+++ src/plugins/pbjx/actions/receiveEnvelope.js
@@ -1,3 +1,8 @@
+import swal from 'sweetalert2';
+import noop from 'lodash/noop';
+import Code from '@gdbots/schemas/gdbots/pbjx/enums/Code';
+import dismissAlert from '@triniti/admin-ui-plugin/actions/dismissAlert';
+import getAlerts from '@triniti/admin-ui-plugin/selectors/getAlerts';
 import { actionTypes } from '../constants';
 
 /**
@@ -15,7 +20,27 @@ import { actionTypes } from '../constants';
  *
  * @returns {Object}
  */
-export default (envelope) => ({
-  type: actionTypes.ENVELOPE_RECEIVED,
-  pbj: envelope,
-});
+export default (envelope) => (dispatch, getState) => {
+  if (envelope.get('code') === Code.UNAUTHENTICATED.getValue()) {
+    setTimeout(() => {
+      getAlerts(getState()).map((alert) => dispatch(dismissAlert(alert.id)));
+      swal.fire({
+        title: 'Session Expired',
+        type: 'error',
+        html: '<p><strong>To avoid losing your work:</strong>'
+          + '<ol class="text-left">'
+          + '<li class="pb-2"><mark><u>DO NOT</u></mark> close or refresh.</li>'
+          + '<li class="pb-2"><a href="/" target="_blank" rel="noopener noreferrer"><strong>Login</strong></a> in a new tab.</li>'
+          + '<li class="pb-2">Once logged in, return to this tab.</li>'
+          + '<li>Click <strong>OK</strong> and then retry your operation.</li>'
+          + '</ol>'
+          + '</p>',
+      }).then(noop).catch(console.error);
+    }, 1000);
+  }
+
+  return {
+    type: actionTypes.ENVELOPE_RECEIVED,
+    pbj: envelope,
+  };
+};