+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
 
+## v0.14.1
+* Add `layout` field to gallery screen.
+* Sort users by first name asc on search users screen.
+* Update viewportMaxHeight css variable for DAM uploader.
+* Notify user when session expired and give instructions on how to prevent losing work.
+
+
 ## v0.14.0
 * Allow users to copy from one tab and paste in another.
 * Display subtitled m3u8 on transcodeable-card